requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
python-telegram-bot==13.7
urllib3==1.26.5
python-dotenv==0.19.0
requests==2.26.0
psycopg2